refactor reflection
