+++ src/components/index.js
@@ -47,6 +47,7 @@ require('./scene/inspector');
 require('./scene/fog');
 require('./scene/keyboard-shortcuts');
 require('./scene/pool');
+require('./scene/reflection');
 require('./scene/screenshot');
 require('./scene/stats');
 require('./scene/vr-mode-ui');
